# api/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Integracion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/publicaciones": {
            "get": {
                "tags": [
                    "Publicaciones"
                ],
                "summary": "Mostrar publicaciones",
                "operationId": "c12dd3fc46f73ab9bd003c1bdd4fe5f3",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las publicaciones."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Publicaciones"
                ],
                "summary": "Crear publicaciones",
                "operationId": "986809601e87e42997c7a68183f447af",
                "responses": {
                    "200": {
                        "description": "Crea la publicación."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/publicaciones/:id": {
            "get": {
                "tags": [
                    "Publicaciones"
                ],
                "summary": "Mostrar una publicacion",
                "operationId": "a5379d5961ad145ed7b025f9b585fbf6",
                "responses": {
                    "200": {
                        "description": "Mostrar una publicacion."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Publicaciones"
                ],
                "summary": "Actualizar publicaciones",
                "operationId": "b3bb76ec3260aee5d1367421b0195c0e",
                "responses": {
                    "200": {
                        "description": "Se ha actualizado el usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/publicacion": {
            "delete": {
                "tags": [
                    "Publicaciones"
                ],
                "summary": "Borra la publicacion",
                "operationId": "bb91b4b5c28788ac4ad24bbb929e9c63",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}